+++ tests/testthat/test-generalized-bootstrap.R
@@ -302,10 +302,11 @@ twophase_design <- twophase(
       })
 
       rescaled_design <- twophase_gen_boot |>
-        rescale_reps(tau = NULL, min_wgt = 0.05)
+        rescale_replicates(min_wgt = 0.05)
       rescaled_matrix <- twophase_gen_boot |>
         weights(type = "replication") |>
-        rescale_reps(tau = NULL, min_wgt = 0.05)
+        `attr<-`('scale', twophase_gen_boot$scale) |>
+          rescale_replicates(min_wgt = 0.05)
 
     expect_equal(
       object = rescaled_design |> weights(type = "replication"),
@@ -318,7 +319,8 @@ twophase_design <- twophase(
 
     expect_equal(
       object = matrix(c(1,0.3,0.02,2,3,4), ncol = 3) |>
-        rescale_reps(min_wgt = 0.02) |>
+        `attr<-`('scale', 1) |>
+        rescale_replicates(min_wgt = 0.02) |>
         `attr<-`('tau', NULL),
       expected = matrix(c(1,0.3,0.02,2,3,4), ncol = 3)
     )